Add startup checks to ensure .env is set. Remove --progress option from webpack-dev-server so it is easier to see errors that the backend throws during startup. Move utils to services/ directory.
