+++ package.json
@@ -5,7 +5,7 @@
   "main": "./public/index.js",
   "scripts": {
     "start": "node src/server.js",
-    "dev": "NODE_ENV=development concurrently \" webpack-dev-server --progress --color \" \"nodemon --inspect src/server.js\"",
+    "dev": "NODE_ENV=development concurrently \" webpack-dev-server --color \" \"nodemon --inspect src/server.js\"",
     "build": "webpack .",
     "test": "echo \"Error: no test specified\" && exit 1",
     "front": "NODE_ENV=development webpack-dev-server --open",
